Update README.md
PAYE Application Program

The program  calculates PAYE taxes based on income, allowances and deductions and finally generates the PAYE certificates/reports. 
.
How to use:
On program execution the user gets:
1.Greetings i.e: "Hello"
2.Prompted to enter their income(Ksh)
3.Their allowances
4.Deductions
5. Relief
The program then outputs the final tax paid using the following rates:

Upto 24000 - 10% tax rate
24001 - 40000 15% tax rate
40001 - 60000 20% tax rate
60001 - 100000 25% tax rate
100001 - 150000 30% tax rate
150001 - 250000 35%  tax rate
Over 250000 40% tax rate.

Finally a certificate is generated for the user.

Error handling:
The program handles errors eg entry of a character or string instead of an integer will output "Invalid" and the user is prompted till the correct output is entered.

The program ensures an easy experience by the user.